Adicionando classes Cliente, Pagamento, Restaurante

+++ src/main/java/com/webapp/agridoce/restaurante/RestauranteRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RestauranteRepository extends JpaRepository<Restaurante, Integer> {
+public interface RestauranteRepository extends JpaRepository<Restaurante, Long> {
 }
